Fix nested subRows selected state not updating properly (#4)
The logic for updating `isSelected` on a row was not being run for all nested subRows. This meant that checking/unchecking a subRow would not update the checked state for the checkbox even though the selectedRowsIds state was correct.

Also change the indeterminate state logic so that a parent row will be considered indeterminate if all direct subRows are also indeterminate (previously would consider indeterminate subRows as unchecked).