Pass subRows as childRows instead of leafRows (#7)
The `childRows` being passed to the aggregate function was always the same as `leafRows`, so it was never useful. Not sure if this was a long-standing bug, but It seems like what we really want is `subRows` from the current grouped level. This `subRows` will be passed as `childRows` now, which allows us to access already aggregated data from leafRows resulting in much better performance.